pc-agent/requirements.txt
websockets==12.0
aiohttp==3.9.1
psutil==5.9.6
pyyaml==6.0.1
python-dotenv==1.0.0

# Windows-only dependencies (install only on Windows)
pywin32==306; sys_platform == 'win32'
